requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
pycpfcnpj==1.8